MEDIUM: sample: Enhances converter "bytes" to take variable names as arguments
Prior to this commit, converter "bytes" takes only integer values as arguments.
After this commit, it can take variable names as inputs.
This allows us to dynamically determine the offset/length and capture them in variables.
These variables can then be used with the converter.
Example use case: parsing a token present in a request header.